feat: Update manifest and dependencies
Update version, description, and host permissions in manifest. Add various dependencies.
