+++ extension/package.json
@@ -4,7 +4,9 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "jest",
+    "test:watch": "jest --watch",
+    "test:coverage": "jest --coverage",
     "dev": "vite",
     "build": "vite build",
     "preview": "vite preview"
@@ -15,8 +17,10 @@
   "private": "true",
   "type": "module",
   "dependencies": {
+    "@vitejs/plugin-react": "^5.0.4",
     "diff-match-patch": "^1.0.5",
     "docx": "^9.5.1",
+    "file-saver": "^2.0.5",
     "html2canvas": "^1.4.1",
     "jspdf": "^3.0.3",
     "mammoth": "^1.11.0",
@@ -27,9 +31,17 @@
   },
   "devDependencies": {
     "@crxjs/vite-plugin": "^2.2.0",
+    "@testing-library/jest-dom": "^6.9.1",
+    "@testing-library/react": "^16.3.0",
+    "@testing-library/user-event": "^14.6.1",
     "@types/chrome": "^0.1.21",
+    "@types/jest": "^30.0.0",
     "@types/react": "^19.2.0",
     "@types/react-dom": "^19.2.0",
+    "identity-obj-proxy": "^3.0.0",
+    "jest": "^30.2.0",
+    "jest-environment-jsdom": "^30.2.0",
+    "ts-jest": "^29.4.4",
     "typescript": "^5.9.3",
     "vite": "^7.1.9"
   }
