+++ extension/src/components/CoverLetter.tsx
