+++ extension/src/components/JobDescriptionInput.tsx
