+++ extension/src/components/ProfileManager.tsx
