+++ extension/src/data/degrees.ts
