+++ extension/src/data/locations.ts
