+++ extension/src/utils/documentGenerator.ts
