--- src/popup.html
@@ -1,20 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>AI CV & Cover Letter Optimizer</title>
-  <style>
-    body {
-      margin: 0;
-      padding: 0;
-      width: 800px;
-      min-height: 600px;
-      font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen', 'Ubuntu', 'Cantarell', sans-serif;
-    }
-  </style>
-</head>
-<body>
-  <div id="root"></div>
-</body>
-</html>
